+++ app/src/main/kotlin/com/fakedevelopers/bidderbidder/ui/product_registration/album_list/AlbumListAdapter.kt
@@ -1,7 +1,6 @@
 package com.fakedevelopers.bidderbidder.ui.product_registration.album_list
 
 import android.content.Context
-import android.graphics.drawable.Drawable
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -9,12 +8,9 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.bumptech.glide.load.DataSource
-import com.bumptech.glide.load.engine.GlideException
-import com.bumptech.glide.request.RequestListener
-import com.bumptech.glide.request.target.Target
 import com.fakedevelopers.bidderbidder.R
 import com.fakedevelopers.bidderbidder.databinding.RecyclerPictureSelectBinding
+import com.fakedevelopers.bidderbidder.ui.util.GlideRequestListener
 
 class AlbumListAdapter(
     private val findSelectedImageIndex: (String) -> Int,
@@ -37,28 +33,10 @@ class AlbumListAdapter(
                     .load(item)
                     .placeholder(R.drawable.the_cat)
                     .error(R.drawable.error_cat)
-                    .listener(object : RequestListener<Drawable> {
-                        override fun onLoadFailed(
-                            e: GlideException?,
-                            model: Any?,
-                            target: Target<Drawable>?,
-                            isFirstResource: Boolean
-                        ): Boolean {
-                            isErrorImage = true
-                            return false
-                        }
-
-                        override fun onResourceReady(
-                            resource: Drawable?,
-                            model: Any?,
-                            target: Target<Drawable>?,
-                            dataSource: DataSource?,
-                            isFirstResource: Boolean
-                        ): Boolean {
-                            isErrorImage = false
-                            return false
-                        }
-                    })
+                    .listener(GlideRequestListener(
+                        loadFailed = { isErrorImage = true },
+                        resourceReady = { isErrorImage = false }
+                    ))
                     .into(image)
             }
             // 선택된 사진 리스트에 현재 item이 포함되어 있다면 표시해줍니다.
