+++ app/src/main/kotlin/com/fakedevelopers/bidderbidder/ui/product_registration/album_list/AlbumPagerAdapter.kt
@@ -1,19 +1,15 @@
 package com.fakedevelopers.bidderbidder.ui.product_registration.album_list
 
 import android.content.Context
-import android.graphics.drawable.Drawable
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.bumptech.glide.load.DataSource
-import com.bumptech.glide.load.engine.GlideException
-import com.bumptech.glide.request.RequestListener
-import com.bumptech.glide.request.target.Target
 import com.fakedevelopers.bidderbidder.R
 import com.fakedevelopers.bidderbidder.databinding.RecyclerAlbumPagerBinding
+import com.fakedevelopers.bidderbidder.ui.util.GlideRequestListener
 
 class AlbumPagerAdapter(
     private val sendErrorToast: () -> Unit,
@@ -30,28 +26,10 @@ class AlbumPagerAdapter(
                 .load(item)
                 .placeholder(R.drawable.the_cat)
                 .error(R.drawable.error_cat)
-                .listener(object : RequestListener<Drawable> {
-                    override fun onLoadFailed(
-                        e: GlideException?,
-                        model: Any?,
-                        target: Target<Drawable>?,
-                        isFirstResource: Boolean
-                    ): Boolean {
-                        isErrorImage = true
-                        return false
-                    }
-
-                    override fun onResourceReady(
-                        resource: Drawable?,
-                        model: Any?,
-                        target: Target<Drawable>?,
-                        dataSource: DataSource?,
-                        isFirstResource: Boolean
-                    ): Boolean {
-                        isErrorImage = false
-                        return false
-                    }
-                })
+                .listener(GlideRequestListener(
+                    loadFailed = { isErrorImage = true },
+                    resourceReady = { isErrorImage = false }
+                ))
                 .into(binding.imageviewAlbumPager)
             binding.layoutAlbumPager.setOnClickListener {
                 if (!isErrorImage) {
