+++ app/src/main/kotlin/com/fakedevelopers/bidderbidder/ui/util/GlideRequestListener.kt
@@ -1,0 +1,34 @@
+package com.fakedevelopers.bidderbidder.ui.util
+
+import android.graphics.drawable.Drawable
+import com.bumptech.glide.load.DataSource
+import com.bumptech.glide.load.engine.GlideException
+import com.bumptech.glide.request.RequestListener
+import com.bumptech.glide.request.target.Target
+
+class GlideRequestListener(
+    private val loadFailed: (() -> Unit)? = null,
+    private val resourceReady: (() -> Unit)? = null
+) : RequestListener<Drawable> {
+
+    override fun onLoadFailed(
+        e: GlideException?,
+        model: Any?,
+        target: Target<Drawable>?,
+        isFirstResource: Boolean
+    ): Boolean {
+        loadFailed?.invoke()
+        return false
+    }
+
+    override fun onResourceReady(
+        resource: Drawable?,
+        model: Any?,
+        target: Target<Drawable>?,
+        dataSource: DataSource?,
+        isFirstResource: Boolean
+    ): Boolean {
+        resourceReady?.invoke()
+        return false
+    }
+}
